Add the ability to create n1ql indexes during tests
This change allows us to create primary and secondary indexes
based on parameters set in the test config files. Two new parameters
are added and one existing perameter is re-used. The parameters are
specified below and should be added to the index section of the
config file.

index_type - The type of secondary indexes to create. The index
types can be found in perfrunner.workloads.n1ql.py
primary_index - Boolean value specifying whether or no the test
should automatically create a promary index. Defaults to false.
use_gsi - Boolean value specifying whether or not gsi index should
be used. Defaults to false.

Lastly, this change adds a basic N1QL test. The test currently
creates a cluster and indexes in the cluster. After this setup
phase is complete the test loads items and runs a workload for
60 seconds.

Change-Id: I2d6bef3e7f6b8c9b2073e524ab8fa22493fc3077