Add a n1ql_stale parameter for stale=false queries
We need to add in a paraemter to modify queries to be able to set
stale=false. This change adds the parameter and update the n1ql
configuration files appropriately. I added stale=ok to some queries
even though it is the default in order to improve test configuration
readability.

Change-Id: Ib522be9a38a249297bf02bcda83a751e972294da